[v0.91.7][WP-08][AWS] Implement AWS signal integration in full (Closes #4686)
