+++ adl/src/lib.rs
@@ -98,3 +98,4 @@ pub mod uts;
 pub mod uts_acc_compiler;
 pub mod uts_acc_multi_model_benchmark;
 pub mod uts_conformance;
+pub mod wp08_acip_sns_proof;
